feat: ajustes gerais

+++ src/views/PesquisaLivroView.java
@@ -5,7 +5,9 @@ import java.awt.*;
 
 import models.Database.LivroDatabase;
 import models.Livro.Livro;
+import models.Livro.LivroCategoria;
 
+import javax.swing.table.DefaultTableModel;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.List;
@@ -16,6 +18,15 @@ public class PesquisaLivroView extends JFrame implements ActionListener {
     private JPanel resultadosPanel;
 
     public PesquisaLivroView() {
+        JMenuBar menuBar = new JMenuBar();
+        JMenu menu = new JMenu("Ações");
+        JMenuItem menuItem = new JMenuItem("Adicionar Livro");
+        menuItem.setActionCommand("adicionarLivro");
+        menuItem.addActionListener(this);
+        menu.add(menuItem);
+        menuBar.add(menu);
+        setJMenuBar(menuBar);
+
         setTitle("Pesquisar Livros");
         setSize(600, 400);
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
@@ -32,11 +43,6 @@ public class PesquisaLivroView extends JFrame implements ActionListener {
         searchPanel.add(textFieldPesquisa, BorderLayout.CENTER);
         searchPanel.add(btnPesquisar, BorderLayout.EAST);
 
-        JButton btnMais = new JButton("Adicionar Livro");
-        btnMais.setActionCommand("adicionarLivro");
-        btnMais.addActionListener(this);
-        searchPanel.add(btnMais, BorderLayout.NORTH);
-
         resultadosPanel = new JPanel();
         resultadosPanel.setLayout(new FlowLayout(FlowLayout.LEFT, 5, 10));
         resultadosPanel.setAlignmentX(Component.LEFT_ALIGNMENT);
@@ -49,35 +55,73 @@ public class PesquisaLivroView extends JFrame implements ActionListener {
 
     private void pesquisarLivros() {
         String pesquisa = textFieldPesquisa.getText();
-
+        List<Livro> livrosEncontrados = LivroDatabase.pesquisarLivro(pesquisa, pesquisa, pesquisa, pesquisa);
         JPanel novoPanel = new JPanel();
         novoPanel.setLayout(new BoxLayout(novoPanel, BoxLayout.Y_AXIS));
 
-        List<Livro> livrosEncontrados = LivroDatabase.pesquisarLivro(pesquisa, pesquisa, pesquisa, pesquisa);
+        ImageIcon lapisIcon = new ImageIcon(getClass().getResource("/media/lapis.png"));
+        ImageIcon lixeiraIcon = new ImageIcon(getClass().getResource("/media/lixeira.png"));
+
+        JPanel escritaPanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
+        JLabel lblEscritaTitulo = new JLabel("Título");
+        JLabel lblEscritaCategoria = new JLabel("Categoria");
+        JLabel lblEscritaAutor = new JLabel("Autor");
+        JLabel lblEscritaEditor = new JLabel("Editora");
+        JLabel lblEscritaAcoes = new JLabel("Ações");
+
+        lblEscritaTitulo.setPreferredSize(new Dimension(100, 20));
+        lblEscritaCategoria.setPreferredSize(new Dimension(100, 20));
+        lblEscritaAutor.setPreferredSize(new Dimension(100, 20));
+        lblEscritaEditor.setPreferredSize(new Dimension(100, 20));
+        lblEscritaAcoes.setPreferredSize(new Dimension(100, 20));
+
+        escritaPanel.add(lblEscritaTitulo);
+        escritaPanel.add(lblEscritaCategoria);
+        escritaPanel.add(lblEscritaAutor);
+        escritaPanel.add(lblEscritaEditor);
+        escritaPanel.add(lblEscritaAcoes);
+        escritaPanel.add(Box.createRigidArea(new Dimension(0, 5)));
+
+        novoPanel.add(escritaPanel);
 
         for (Livro livro : livrosEncontrados) {
-            JPanel livroPanel = new JPanel(new BorderLayout());
+            JPanel livroPanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
             JLabel lblTitulo = new JLabel(livro.getTitulo());
+            JLabel lblCategoria = new JLabel(LivroCategoria.getDescricao(livro.getCategoria()));
+            JLabel lblAutor = new JLabel(livro.getAutor());
+            JLabel lblEditora = new JLabel(livro.getEditora());
+
+            lblTitulo.setPreferredSize(new Dimension(100, 15));
+            lblCategoria.setPreferredSize(new Dimension(100, 15));
+            lblAutor.setPreferredSize(new Dimension(100, 15));
+            lblEditora.setPreferredSize(new Dimension(100, 15));
+
+            livroPanel.add(lblTitulo);
+            livroPanel.add(lblCategoria);
+            livroPanel.add(lblAutor);
+            livroPanel.add(lblEditora);
 
             JPanel btnPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-            JButton btnEditar = new JButton(new ImageIcon("media/lapis.png"));
-            JButton btnExcluir = new JButton(new ImageIcon("media/lixeira.png"));
+
+            JButton btnEditar = new JButton(lapisIcon);
+            JButton btnExcluir = new JButton(lixeiraIcon);
 
             btnEditar.setActionCommand("editar:" + livro.getIdLivro());
             btnExcluir.setActionCommand("excluir:" + livro.getIdLivro());
 
             btnEditar.setPreferredSize(new Dimension(20, 20));
             btnExcluir.setPreferredSize(new Dimension(20, 20));
+
             btnEditar.addActionListener(this);
             btnExcluir.addActionListener(this);
 
             btnPanel.add(btnEditar);
             btnPanel.add(btnExcluir);
 
-            livroPanel.add(lblTitulo, BorderLayout.CENTER);
-            livroPanel.add(btnPanel, BorderLayout.WEST);
-
+            livroPanel.add(btnPanel);
             novoPanel.add(livroPanel);
+
+            novoPanel.add(Box.createRigidArea(new Dimension(0, 1)));
         }
 
         resultadosPanel.removeAll();
@@ -115,8 +159,13 @@ public class PesquisaLivroView extends JFrame implements ActionListener {
                 } else if (e.getActionCommand().startsWith("excluir:")) {
                     String idLivroStr = e.getActionCommand().substring("excluir:".length());
                     int idLivroExcluir = Integer.parseInt(idLivroStr);
-                    LivroDatabase.excluirLivro(idLivroExcluir);
-                    pesquisarLivros();
+
+                    int opcao = JOptionPane.showConfirmDialog(null, "Tem certeza que deseja excluir?", "Confirmação", JOptionPane.YES_NO_OPTION);
+
+                    if (opcao == JOptionPane.YES_OPTION) {
+                        LivroDatabase.excluirLivro(idLivroExcluir);
+                        pesquisarLivros();
+                    }
                 }
         }
     }
